graphics/python/requirements.txt
Pillow
shapely
numpy
matplotlib
natsort
imageio
opencv-python
